rename name to title
