+++ packages/compass-aggregations/src/plugin-title.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
 
-export function AggregationsPluginName() {
+export function AggregationsTabTitle() {
   return <div data-testid="aggregations-tab-title">Aggregations</div>;
 }
