+++ packages/compass-indexes/src/plugin-title.tsx
@@ -1,7 +1,13 @@
 import React, { useMemo } from 'react';
+import { connect } from 'react-redux';
+import type { RootState } from './modules';
+import { Badge, css, spacing, Tooltip } from '@mongodb-js/compass-components';
 import numeral from 'numeral';
-import { css, Tooltip, Badge, spacing } from '@mongodb-js/compass-components';
-import type { CrudStore } from './stores/crud-store';
+
+const containerStyles = css({
+  display: 'flex',
+  gap: spacing[200],
+});
 
 const tooltipContentStyles = css({
   listStyleType: 'none',
@@ -9,12 +15,14 @@ const tooltipContentStyles = css({
   margin: 0,
 });
 
-const containerStyles = css({
-  display: 'flex',
-  gap: spacing[200],
-});
-
 const INVALID = 'N/A';
+
+const avg = (size: number, count: number) => {
+  if (count <= 0) {
+    return 0;
+  }
+  return size / count;
+};
 
 const isNumber = (val: any): val is number => {
   return typeof val === 'number' && !isNaN(val);
@@ -67,36 +75,34 @@ const CollectionStats: React.FunctionComponent<CollectionStatsProps> = ({
   );
 };
 
-export const CrudPluginName = ({
-  store: {
-    state: { collectionStats },
-  },
+const TabTitle = ({
+  collectionStats,
 }: {
-  store: CrudStore;
+  collectionStats: RootState['collectionStats'];
 }) => {
-  const { documentCount, storageSize, avgDocumentSize } = useMemo(() => {
-    const {
-      document_count = NaN,
-      storage_size = NaN,
-      free_storage_size = NaN,
-      avg_document_size = NaN,
-    } = collectionStats ?? {};
+  const { indexCount, totalIndexSize, avgIndexSize } = useMemo(() => {
+    const { index_count = NaN, index_size = NaN } = collectionStats ?? {};
     return {
-      documentCount: format(document_count),
-      storageSize: format(storage_size - free_storage_size, 'b'),
-      avgDocumentSize: format(avg_document_size, 'b'),
+      indexCount: format(index_count),
+      totalIndexSize: format(index_size, 'b'),
+      avgIndexSize: format(avg(index_size, index_count), 'b'),
     };
   }, [collectionStats]);
+
   const details = [
-    `Documents: ${documentCount}`,
-    `Storage Size: ${storageSize}`,
-    `Avg. Size: ${avgDocumentSize}`,
+    `Indexes: ${indexCount}`,
+    `Total Size: ${totalIndexSize}`,
+    `Avg. Size: ${avgIndexSize}`,
   ];
 
   return (
-    <div data-testid="documents-tab-title" className={containerStyles}>
-      Documents
-      <CollectionStats text={documentCount} details={details} />
+    <div data-testid="indexes-tab-title" className={containerStyles}>
+      Indexes
+      <CollectionStats text={indexCount} details={details} />
     </div>
   );
 };
+
+export const IndexesTabTitle = connect(({ collectionStats }: RootState) => ({
+  collectionStats,
+}))(TabTitle);
