+++ packages/compass-schema-validation/src/plugin-title.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
 
-export function SchemaValidationPluginName() {
+export function SchemaValidationTabTitle() {
   return <div data-testid="validation-tab-title">Validation</div>;
 }
